+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### 1.0.10
+*
+
 ### 1.0.9
 * Add Apache Felix Jetty Based HTTP Service, Apache Sling GET Servlet, Apache Sling Referrer Filter, Apache Sling DavEx Servlet OSGI config support